requirements.txt
streamlit
pandas
geopandas
folium
streamlit_folium
openai
python-dotenv
fuzzywuzzy
langchain
requests
streamlit_chat
